refactor: extract controller into submodules
Split the monolithic airflow_controller.rs into controller/ with
submodules: dereference, validate, build (+ build/role_group_builder),
apply, and update_status.

This is a purely structural refactor — no logic changes. Each pipeline
stage now lives in its own file for easier navigation and review.

Rename airflow_controller module to controller in main.rs.
