+++ rust/operator-binary/src/main.rs
@@ -33,12 +33,12 @@ use stackable_operator::{
 };
 
 use crate::{
-    airflow_controller::AIRFLOW_FULL_CONTROLLER_NAME,
+    controller::AIRFLOW_FULL_CONTROLLER_NAME,
     crd::{AirflowCluster, AirflowClusterVersion, OPERATOR_NAME, v1alpha1, v1alpha2},
     webhooks::conversion::create_webhook_server,
 };
 
-mod airflow_controller;
+mod controller;
 mod config;
 mod controller_commons;
 mod crd;
@@ -178,9 +178,9 @@ async fn main() -> anyhow::Result<()> {
                 .graceful_shutdown_on(sigterm_watcher.handle())
                 .run(
                     // REVIEW: renamed from reconcile_airflow to reconcile
-                    airflow_controller::reconcile,
-                    airflow_controller::error_policy,
-                    Arc::new(airflow_controller::Ctx {
+                    controller::reconcile,
+                    controller::error_policy,
+                    Arc::new(controller::Ctx {
                         client: client.clone(),
                         operator_environment,
                         product_config,
